requirements.txt
Transformers
torch
Pillow
fastapi 
uvicorn 
# 💡 Why: These packages are required for FastAPI (API), PyTorch (model backend), Hugging Face Transformers (image captioning), and Pillow (image processing).